add authorization middleware in routes & models

+++ src/db/models/parkingSpot.js
@@ -3,18 +3,28 @@ const database = require("../config/mongodb");
 const { ObjectId } = require("mongodb");
 
 class ParkingSpotModels {
-  static async getAll() {
+  static async getAll({ role, authorId }) {
+    let opt = {};
+    if (role === "vendor") {
+      opt = {
+        authorId,
+      };
+    }
+
     const parkSpots = await database
       .collection("parkingSpots")
-      .find()
+      .find(opt)
       .toArray();
     return parkSpots;
   }
 
-  static async getById(id) {
-    const parkingSpot = await database
-      .collection("parkingSpots")
-      .findOne({ _id: new ObjectId(String(id)) });
+  static async getById({ id, authorId, role }) {
+    let opt = { _id: new ObjectId(String(id)) };
+    if (role === "vendor") {
+      opt["$and"] = [{ authorId: new ObjectId(String(authorId)) }];
+    }
+
+    const parkingSpot = await database.collection("parkingSpots").findOne(opt);
     if (!parkingSpot) throw { name: "ParkingSpotNotFound" };
     return parkingSpot;
   }
@@ -27,7 +37,11 @@ class ParkingSpotModels {
     carSpot,
     motorFee,
     carFee,
+    authorId,
+    role,
   }) {
+    if (role !== "vendor")
+      throw { name: "Your account unauthorized to create parking spot" };
     const validation = z
     .object({
       name: z.string().min(1, "is required"),
@@ -54,11 +68,11 @@ class ParkingSpotModels {
       carSpot,
       motorFee,
       carFee,
-      authorId: new ObjectId(String("66d6d3d0cf201705437e09cc")),
+      authorId: new ObjectId(String(authorId)),
     });
     return "Success create parking spot";
   }
-  
+
   static async updateParkingSpot({
     id,
     name,
@@ -68,7 +82,10 @@ class ParkingSpotModels {
     carSpot,
     motorFee,
     carFee,
+    role,
   }) {
+    if (role !== "vendor")
+      throw { name: "Your account unauthorized to update parking spot data" };
     const validation = z
     .object({
       name: z.string().min(1, "is required"),
@@ -96,13 +113,15 @@ class ParkingSpotModels {
     return "Success update parking spot";
   }
 
-
-  static async deleteParkingSpot(id) {
-    const result = await database.collection("parkingSpots").deleteOne(
-      {
-        _id: new ObjectId(String(id)),
-      }
-    )
+  static async deleteParkingSpot({ id, authorId, role }) {
+    if (role !== "vendor")
+      throw { name: "Your account unauthorized to delete parking spot" };
+    const result = await database.collection("parkingSpots").deleteOne({
+      $and: [
+        { _id: new ObjectId(String(id)) },
+        { authorId: new ObjectId(String(authorId)) },
+      ],
+    });
     return { result: "Success delete parking spot" };
   }
 }
