+++ src/db/models/user.js
@@ -80,7 +80,8 @@ class UserModels {
         return { user: "success create user" }
     }
 
-    static async addVendor({ name, username, email, password }) {
+    static async addVendor({ name, username, email, password, userRole }) {
+        if(userRole !== "admin") throw { name: "Unauthorized" }
         const validation = z
             .object({
                 name: z.string().min(1, "is required"),
@@ -103,7 +104,8 @@ class UserModels {
         return { user: "success create vendor" }
     }
 
-    static async getAll() {
+    static async getAll({role}) {
+        if(role !== "admin") throw { name: "Unauthorized" }
         const agg = [
             {
                 '$project': {
@@ -178,7 +180,8 @@ class UserModels {
         return { result: "Success update" };
     }
 
-    static async deleteUser(id) {
+    static async deleteUser({id, role}) {
+        if (role !== "admin") throw { name: "Unauthorized" }
         const result = await database.collection("users").deleteOne(
             {
                 _id: new ObjectId(String(id)),
